Create and add fragment to activity
