+++ app/src/main/res/layout/list_item.xml
@@ -1,7 +1,7 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
-    android:layout_height="wrap_content" >
+    android:layout_height="wrap_content">
 
     <ImageView
         android:id="@+id/file_icon"
@@ -14,7 +14,7 @@
         android:id="@+id/file_name"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="@string/file_name"
         android:layout_margin="@dimen/layout_margin"
+        android:text="@string/file_name"
         android:textAppearance="?android:attr/textAppearanceLarge" />
 </LinearLayout>
