+++ app/src/main/res/values/strings.xml
@@ -1,8 +1,8 @@
 <resources>
     <string name="app_name">"File Manager"</string>
+    <string name="chooser_title">Choose application</string>
     <string name="empty_folder">Folder is empty :(</string>
     <string name="file_name">File name</string>
     <string name="file_path">File path</string>
-    <string name="missing_file_handler">No handler for this type of file.</string>
-    <string name="chooser_title">Choose application</string>
+    <string name="missing_file_handler">No handler for this type of file</string>
 </resources>
